factor out tests of first-resolved and first-value
